+++ apps/shelve/app/app.vue
@@ -1,13 +1,6 @@
 <script setup lang="ts">
 import { Toaster } from 'vue-sonner'
 
-useScriptPlausibleAnalytics({
-  domain: 'shelve.cloud',
-  scriptInput: {
-    src: 'https://analytics.hrcd.fr/js/script.js',
-  }
-})
-
 const { title, link, description, ogImage } = useAppConfig()
 
 useHead({
