--- apps/shelve/app/components/landing/Sparkles.client.vue
@@ -1,123 +1,0 @@
-<script setup lang="ts">
-import { tsParticles } from '@tsparticles/engine'
-import { loadSlim } from '@tsparticles/slim'
-import { onMounted, onUnmounted } from 'vue'
-
-const props = defineProps({
-  id: {
-    type: String,
-    default: () => `sparkles-${Math.round(Math.random() * 9999)}`,
-  },
-  size: {
-    type: Number,
-    default: 1,
-  },
-  minSize: {
-    type: Number,
-    default: 0.4,
-  },
-  density: {
-    type: Number,
-    default: 800,
-  },
-  speed: {
-    type: Number,
-    default: 1,
-  },
-  minSpeed: {
-    type: Number,
-    required: false,
-  },
-  opacity: {
-    type: Number,
-    default: 1,
-  },
-  opacitySpeed: {
-    type: Number,
-    default: 3,
-  },
-  color: {
-    type: String,
-    default: '#FFFFFF',
-  },
-  background: {
-    type: String,
-    default: 'transparent',
-  },
-  options: {
-    type: Object,
-    required: false,
-  },
-})
-
-const defaultOptions = {
-  background: {
-    color: {
-      value: props.background,
-    },
-  },
-  fullScreen: {
-    enable: false,
-    zIndex: 1,
-  },
-  fpsLimit: 120,
-  particles: {
-    color: {
-      value: props.color,
-    },
-    move: {
-      enable: true,
-      direction: 'none',
-      speed: {
-        min: props.minSpeed || props.speed / 10,
-        max: props.speed,
-      },
-      straight: false,
-    },
-    number: {
-      value: props.density,
-    },
-    opacity: {
-      value: {
-        min: props.minOpacity || props.opacity / 10,
-        max: props.opacity,
-      },
-      animation: {
-        enable: true,
-        sync: false,
-        speed: props.opacitySpeed,
-      },
-    },
-    size: {
-      value: {
-        min: props.minSize || props.size / 2.5,
-        max: props.size,
-      },
-    },
-  },
-  detectRetina: true,
-}
-
-let container
-
-onMounted(async () => {
-  await loadSlim(tsParticles)
-
-  container = await tsParticles.load({
-    id: props.id,
-    options: props.options || defaultOptions,
-  })
-})
-
-onUnmounted(() => {
-  if (!container) {
-    return
-  }
-
-  container.destroy()
-})
-</script>
-
-<template>
-  <div :id />
-</template>
