+++ apps/shelve/app/middleware/project-redirect.ts
@@ -1,6 +1,6 @@
 export default defineNuxtRouteMiddleware(((to, from) => {
   const projectId = to.params.projectId || from.params.projectId || ''
-  if (to.path === `/app/project/${projectId}`) {
-    return `/app/project/${projectId}/variables`
+  if (to.path === `/project/${projectId}`) {
+    return `/project/${projectId}/variables`
   }
 }))
