+++ apps/shelve/app/utils/navigation.ts
@@ -1,4 +1,4 @@
-type Where = 'home' | 'app'| 'admin';
+type Where = 'app'| 'admin';
 
 export type Navigation = {
   name: string;
@@ -9,43 +9,35 @@ export type Navigation = {
 
 export function getNavigation(where: Where): Navigation[] {
   switch (where) {
-    case 'home':
-      return [
-        { name: 'Home', path: '/', title: 'Home' },
-        { name: 'Vault', path: '/vault', title: 'Vault' },
-        /*{ name: 'Changelog', path: '/changelog', icon: 'i-heroicons-document-text', title: 'Changelog' },*/
-        { name: 'Docs', path: '/docs', title: 'Docs' },
-        { name: 'Roadmap', path: '/roadmap', title: 'Roadmap' },
-      ]
     case 'app':
       return [
         {
           name: 'Projects',
-          path: '/app/projects',
+          path: '/',
           icon: 'lucide:folder',
           title: 'Projects',
         },
         {
           name: 'Teams',
-          path: '/app/teams',
+          path: '/teams',
           icon: 'lucide:users',
           title: 'Teams',
         },
         {
           name: 'Profile',
-          path: '/app/profile',
+          path: '/profile',
           icon: 'heroicons:user-circle',
           title: 'Profile',
         },
         {
           name: 'Tokens',
-          path: '/app/tokens',
+          path: '/tokens',
           icon: 'heroicons:key',
           title: 'Tokens',
         },
         {
           name: 'Settings',
-          path: '/app/settings',
+          path: '/settings',
           icon: 'heroicons:cog',
           title: 'Settings',
         },
@@ -54,13 +46,13 @@ export function getNavigation(where: Where): Navigation[] {
       return [
         {
           name: 'Users',
-          path: '/app/admin/users',
+          path: '/admin/users',
           icon: 'heroicons:users',
           title: 'Users',
         },
         {
           name: 'Tests',
-          path: '/app/admin/tests',
+          path: '/admin/tests',
           icon: 'lucide:flask-conical',
           title: 'Tests',
         },
