--- apps/shelve/content/release/1.shelve-is-public.md
@@ -1,18 +1,0 @@
----
-title: Shelve is now public
-date: 01-04-2024
-image: /release/shelve-is-public.webp
-tag: public
-authors:
-  - HugoRCD
----
-
-# Announcing the public release of Shelve
-
-We are excited to announce that Shelve is now public! 🎉
-
-## Support
-If you like this project, you can support me by giving a ⭐️ on this Github repository.
-
-## Credits
-- [Hugo Richard](https://x.com/HugoRCD__)
