+++ apps/shelve/nuxt.config.ts
@@ -18,17 +18,10 @@ export default defineNuxtConfig({
     componentIslands: true,
   },
 
-  routeRules: {
-    '/': { isr: true, prerender: true },
-    '/vault': { isr: true, prerender: true },
-    '/app/**': { ssr: false, robots: false },
-  },
+  ssr: false,
 
   nitro: {
     preset: process.env.NITRO_PRESET || 'bun',
-    prerender: {
-      routes: ['/sitemap.xml']
-    },
     rollupConfig: {
       // @ts-expect-error - vue is an external plugin
       plugins: [vue()]
@@ -62,21 +55,8 @@ export default defineNuxtConfig({
     '@vueuse/nuxt',
     'nuxt-build-cache',
     'nuxt-auth-utils',
-    '@nuxt/scripts',
     '@nuxtjs/seo',
   ],
-
-  site: {
-    url: process.env.NUXT_PUBLIC_SITE_URL,
-    name: 'Shelve',
-    description: 'Shelve is a project management tool for developers teams',
-    defaultLocale: 'en',
-    indexable: true,
-  },
-
-  sitemap: {
-    exclude: ['/app/**'],
-  },
 
   css: ['~/assets/style/main.css'],
