+++ apps/shelve/server/routes/auth/github.ts
@@ -28,7 +28,7 @@ export default defineOAuthGitHubEventHandler({
         },
         loggedInAt: new Date().toISOString(),
       })
-      return sendRedirect(event, '/app/projects')
+      return sendRedirect(event, '/')
     } catch (error) {
       console.error('GitHub OAuth error:', error)
       return sendRedirect(event, '/login?error=github')
